apps/ingestor/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
google-cloud-pubsub==2.21.6
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
pydantic==2.8.2